refactores format document
